Changelog: New cmath Module, Performance Optimizations, and Enhanced Stability in ZzzMathLib
New module: Introduction of cmath for performing advanced mathematical calculations.
Performance optimizations: Significant improvements in the performance of ZzzMaNew module: Introduction of cmath for performing advanced mathematical calculations.
Performance optimizations: Significant improvements in the performance of ZzzMathLib.
Enhancements in overall stability: Refinements aimed at ensuring greater stability in the functioning of ZzzMathLib.cements in overall stability: Refinements aimed at ensuring greater stability in the functioning of ZzzMathLib.

+++ source/GMEmpireGuardian1.cpp
@@ -2039,7 +2039,7 @@ bool GMEmpireGuardian1::RenderMonsterVisual(CHARACTER* c, OBJECT* o, BMD* b)
 		{
 			vec3_t vRelative;
 			Vector( 0.0f, 0.0f ,0.0f ,vRelative )
-			float fLumi1 = (sinf(WorldTime*0.004f) + 0.9f) * 0.25f;
+			float fLumi1 = (std::sin(WorldTime * 0.004f) + 0.9f) * 0.25f;
 			
 			Vector(0.05f+fLumi1, 0.75f+fLumi1, 0.35f+fLumi1, vLight);
